+++ src/extension.ts
@@ -1,5 +1,5 @@
 import * as vscode from 'vscode'
-import { checkSplitPanels, calculateRange } from './utils'
+import { checkSplitPanels, calculateRange, wholeLine, calculatePosition } from './utils'
 import { OnOffState, ModeState, AllStates } from './states'
 
 export function activate(context: vscode.ExtensionContext) {
@@ -27,12 +27,37 @@ export function activate(context: vscode.ExtensionContext) {
 		modeState.registerCommand(() => {
 			reset()
 		}),
+		vscode.commands.registerTextEditorCommand('syncScroll.jumpToNextPanelCorrespondingPosition', (textEditor) => {
+			const selection = textEditor.selection
+			const textEditors = vscode.window.visibleTextEditors
+			.filter(editor => editor !== textEditor && editor.document.uri.scheme !== 'output')
+			const nextTextEditor = textEditors[(textEditors.indexOf(textEditor) + 1) % textEditors.length]
+			const offset = offsetByEditors.get(nextTextEditor)
+			const correspondingStartPosition = calculatePosition(selection.start, offset, textEditor, nextTextEditor)
+			const correspondingPosition = new vscode.Range(correspondingStartPosition, correspondingStartPosition)
+			const correspondingRange = calculateRange(selection, offset)
+			vscode.window.showTextDocument(nextTextEditor.document, {
+				viewColumn: nextTextEditor.viewColumn,
+				selection: selection.isEmpty ? correspondingPosition : correspondingRange
+			})
+		}),
+		vscode.commands.registerTextEditorCommand('syncScroll.copyToAllCorrespondingPositions', (textEditor) => {
+			vscode.window.visibleTextEditors
+				.filter(editor => editor !== textEditor && editor.document.uri.scheme !== 'output')
+				.forEach(scrolledEditor => {
+					scrolledEditor.edit(editBuilder =>
+						textEditor.selections.map(selection =>
+							editBuilder.replace(
+								calculateRange(selection, offsetByEditors.get(scrolledEditor)),
+								textEditor.document.getText(selection.isEmpty ? wholeLine(selection) : selection) + '\n')))
+				})
+		}),
 		vscode.window.onDidChangeVisibleTextEditors(textEditors => {
 			AllStates.areVisible = checkSplitPanels(textEditors)
 			reset()
 		}),
 		vscode.window.onDidChangeTextEditorVisibleRanges(({ textEditor, visibleRanges }) => {
-			if (!AllStates.areVisible || onOffState.isOff() || textEditor.viewColumn === undefined) {
+			if (!AllStates.areVisible || onOffState.isOff() || textEditor.viewColumn === undefined || textEditor.document.uri.scheme === 'output') {
 				return
 			}
 			if (scrollingEditor !== textEditor) {
@@ -43,7 +68,7 @@ export function activate(context: vscode.ExtensionContext) {
 				scrollingEditor = textEditor
 				if (modeState.isOffsetMode()) {
 					vscode.window.visibleTextEditors
-						.filter(editor => editor !== textEditor)
+						.filter(editor => editor !== textEditor && editor.document.uri.scheme !== 'output')
 						.forEach(scrolledEditor => {
 							offsetByEditors.set(scrolledEditor, scrolledEditor.visibleRanges[0].start.line - textEditor.visibleRanges[0].start.line)
 						})
@@ -56,7 +81,7 @@ export function activate(context: vscode.ExtensionContext) {
 			}
 			scrollingTask = setTimeout(() => {
 				vscode.window.visibleTextEditors
-					.filter(editor => editor !== textEditor)
+					.filter(editor => editor !== textEditor && editor.document.uri.scheme !== 'output')
 					.forEach(scrolledEditor => {
 						scrolledEditorsQueue.add(scrolledEditor)
 						scrolledEditor.revealRange(
@@ -67,13 +92,13 @@ export function activate(context: vscode.ExtensionContext) {
 			}, 0)
 		}),
 		vscode.window.onDidChangeTextEditorSelection(({ selections, textEditor }) => {
-			if (!AllStates.areVisible || onOffState.isOff() || textEditor.viewColumn === undefined) {
+			if (!AllStates.areVisible || onOffState.isOff() || textEditor.viewColumn === undefined || textEditor.document.uri.scheme === 'output') {
 				return
 			}
 			correspondingLinesHighlight?.dispose()
 			correspondingLinesHighlight = vscode.window.createTextEditorDecorationType({ backgroundColor: new vscode.ThemeColor('editor.inactiveSelectionBackground') })
 			vscode.window.visibleTextEditors
-				.filter(editor => editor !== textEditor)
+				.filter(editor => editor !== textEditor && editor.document.uri.scheme !== 'output')
 				.forEach((scrolledEditor) => {
 					scrolledEditor.setDecorations(
 						correspondingLinesHighlight!,