+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to the "sync-scroll" extension will be documented in this fi
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+### [1.3.0]
+
+- Add command to jump to corresponding position in the next panel
+- Add command to copy selections to all corresponding positions.
+- Fix the issue of the output panel which shouldn't be involved in the scrolling sync.
+
 ### [1.2.0]
 
 - Add corresponding line highlight feature.